pyproject.toml
[tool.poetry]
name = "vehicle-price-prediction"
version = "1.0.0"
description = "Advanced vehicle price prediction system with real-time predictions and explainable AI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
numpy = "^1.24.3"
scikit-learn = "^1.3.2"
xgboost = "^2.0.3"
lightgbm = "^4.1.0"
catboost = "^1.2.2"
tensorflow = "^2.15.0"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
prometheus-client = "^0.19.0"
psutil = "^5.9.6"
loguru = "^0.7.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
joblib = "^1.3.2"
tqdm = "^4.66.1"
mlflow = "^2.8.1"
dvc = "^3.27.0"
shap = "^0.42.1"
lime = "^0.2.0.1"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"
notebook = "^7.0.6"
safety = "^2.3.5"
bandit = "^1.7.5"
codecov = "^2.1.13"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vehicle-price-predictor = "main:main"
devrun = "main:dev_run"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "tensorflow.*",
    "xgboost.*",
    "lightgbm.*",
    "catboost.*",
    "sklearn.*",
    "prometheus_client.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests"
]
